add more tests and change folder management
